cleanup efficientnet
