+++ composer/models/resnet20_cifar10/model.py
@@ -14,7 +14,7 @@ __all__ = ["CIFAR10_ResNet20"]
 class CIFAR10_ResNet20(ComposerClassifier):
     """A ResNet-20 model extending :class:`.ComposerClassifier`.
 
-    From `Deep Residual Learning for Image Recognition <https://arxiv.org/abs/1512.03385>`_ (He et al, 2016).
+    From `Deep Residual Learning for Image Recognition <https://arxiv.org/abs/1512.03385>`_.
 
     Args:
         num_classes (int, optional): The number of classes. Needed for classification tasks. Default: ``10``.
