+++ composer/models/resnet9_cifar10/model.py
@@ -19,7 +19,7 @@ class ResNet9(nn.Module):
     """A 9-layer residual network, excluding BatchNorms and activation functions, as described in the
     `myrtle.ai blog <https://myrtle.ai/learn/how-to-train-your-resnet-4-architecture/>`_.
 
-    See `Deep Residual Learning for Image Recognition <https://arxiv.org/abs/1512.03385>`_ (He et al, 2016) for details on the residual network architecture.
+    See `Deep Residual Learning for Image Recognition <https://arxiv.org/abs/1512.03385>`_ for details on the residual network architecture.
 
     Args:
         num_classes (int, optional): The number of classes. Needed for classification tasks. Default: ``10``.
